refactor(huspacy): code cleanup, components package moved one level up
