+++ hu_core_news_lg/project.yml
@@ -350,9 +350,9 @@ commands:
     help: "Train the lookup lemmatizer on the full Szeged Corpus"
     script:
       - "mkdir -p ${vars.models_path}/${vars.package_name}-lookup-lemmatizer-${vars.package_version}"
-      - "python ../scripts/train_lookup_lemmatizer.py 
+      - bash -c "PYTHONPATH=\"../huspacy\" python ../scripts/train_lookup_lemmatizer.py 
             ${vars.processed_data_path}/${vars.szegedcorpus}/train.conllu
-            ${vars.models_path}/${vars.package_name}-lookup-lemmatizer-${vars.package_version}"
+            ${vars.transformer_models_path}/${vars.package_name}-lookup-lemmatizer-${vars.package_version}"
     deps:
       - ${vars.processed_data_path}/${vars.szegedcorpus}/train.conllu
     outputs:
