+++ huspacy/huspacy/components/edit_tree_lemmatizer.py
