+++ huspacy/huspacy/components/hun_sentencizer.py
