--- huspacy/tools/custom_code.py
@@ -1,7 +1,0 @@
-import sys
-sys.path.append("../huspacy/tools")
-
-import components.lemma_postprocessing
-import components.lookup_lemmatizer
-import components.edit_tree_lemmatizer
-import components.hun_sentencizer
